% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_info}
\alias{.format_info}
\title{Format dataset info for display}
\usage{
.format_info(info)
}
\arguments{
\item{info}{List containing dataset info}
}
\value{
Formatted string
}
\description{
Format dataset info for display
}
\keyword{internal}
