% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{.get_datasets}
\alias{.get_datasets}
\title{Get all datasets for a modality}
\usage{
.get_datasets(modality)
}
\arguments{
\item{modality}{Modality name}
}
\value{
Named list of dataset entries
}
\description{
Get all datasets for a modality
}
\keyword{internal}
