% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{.get_entry}
\alias{.get_entry}
\title{Get a specific dataset entry}
\usage{
.get_entry(modality, name)
}
\arguments{
\item{modality}{Modality name}

\item{name}{Dataset name}
}
\value{
Dataset entry list
}
\description{
Get a specific dataset entry
}
\keyword{internal}
