% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{.get_modalities}
\alias{.get_modalities}
\title{Get all available modalities}
\usage{
.get_modalities()
}
\value{
Character vector of modality names
}
\description{
Get all available modalities
}
\keyword{internal}
