% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.load_data}
\alias{.load_data}
\title{Load data from a package}
\usage{
.load_data(pkg, dataset = NULL, loader = NULL)
}
\arguments{
\item{pkg}{Package name}

\item{dataset}{Dataset name within the package}

\item{loader}{Optional custom loader function}
}
\value{
The loaded data object
}
\description{
Load data from a package
}
\keyword{internal}
