% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.require_pkg}
\alias{.require_pkg}
\title{Require a package, installing if necessary}
\usage{
.require_pkg(pkg, bioc = TRUE)
}
\arguments{
\item{pkg}{Package name}

\item{bioc}{Whether this is a Bioconductor package}
}
\value{
TRUE if successful
}
\description{
Require a package, installing if necessary
}
\keyword{internal}
