% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.R
\name{tiny_atac}
\alias{tiny_atac}
\title{Load ATAC-seq datasets}
\usage{
tiny_atac(name = "tcga_brca", verbose = TRUE)
}
\arguments{
\item{name}{Dataset name}

\item{verbose}{Print loading info}
}
\value{
SummarizedExperiment or similar
}
\description{
Convenience function for ATAC-seq data.
}
