% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.R
\name{tiny_bulk}
\alias{tiny_bulk}
\title{Load bulk RNA-seq datasets}
\usage{
tiny_bulk(name = "airway", verbose = TRUE)
}
\arguments{
\item{name}{Dataset name: "airway", "tcga_brca", "tcga_luad", "tcga_paad"}

\item{verbose}{Print loading info}
}
\value{
RangedSummarizedExperiment or MultiAssayExperiment object
}
\description{
Convenience function for bulk RNA-seq data.
}
\examples{
\dontrun{
airway <- tiny_bulk("airway")
brca <- tiny_bulk("tcga_brca")
}

}
