% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.R
\name{tiny_chipseq}
\alias{tiny_chipseq}
\title{Load ChIP-seq datasets}
\usage{
tiny_chipseq(name = "tamoxifen", verbose = TRUE)
}
\arguments{
\item{name}{Dataset name: "tamoxifen", "tamoxifen_peaks"}

\item{verbose}{Print loading info}
}
\value{
SummarizedExperiment or DBA object
}
\description{
Convenience function for ChIP-seq data.
}
\examples{
\dontrun{
chip <- tiny_chipseq("tamoxifen")
}

}
