% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.R
\name{tiny_cytof}
\alias{tiny_cytof}
\title{Load CyTOF / mass cytometry datasets}
\usage{
tiny_cytof(name = "levine32", verbose = TRUE)
}
\arguments{
\item{name}{Dataset name: "levine32", "levine13", "samusik", "bodenmiller", "krieg"}

\item{verbose}{Print loading info}
}
\value{
SummarizedExperiment object
}
\description{
Convenience function for CyTOF data.
}
\examples{
\dontrun{
cytof <- tiny_cytof("levine32")
bcr <- tiny_cytof("bodenmiller")
}

}
