% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.R
\name{tiny_genotype}
\alias{tiny_genotype}
\title{Load genotype / variant datasets}
\usage{
tiny_genotype(name = "chr22", verbose = TRUE)
}
\arguments{
\item{name}{Dataset name: "chr22", "structural"}

\item{verbose}{Print loading info}
}
\value{
VCF object
}
\description{
Convenience function for VCF/genotype data.
}
\examples{
\dontrun{
vcf <- tiny_genotype("chr22")
}

}
