% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.R
\name{tiny_list}
\alias{tiny_list}
\title{List available datasets}
\usage{
tiny_list(modality = NULL)
}
\arguments{
\item{modality}{Optional: show datasets for a specific modality only.}
}
\value{
Invisibly returns a data.frame of datasets. Also prints to console.
}
\description{
Show available modalities and datasets.
}
\examples{
\dontrun{
tiny_list()              # All datasets
tiny_list("scrna")       # Just scRNA-seq
tiny_list("spatial")     # Just spatial
}

}
