% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.R
\name{tiny_multiomics}
\alias{tiny_multiomics}
\title{Load multi-omics datasets}
\usage{
tiny_multiomics(name = "tcga_brca_multi", verbose = TRUE)
}
\arguments{
\item{name}{Dataset name: "tcga_brca_multi", "tcga_acc_multi"}

\item{verbose}{Print loading info}
}
\value{
MultiAssayExperiment object
}
\description{
Convenience function for multi-omics data.
}
\examples{
\dontrun{
mae <- tiny_multiomics("tcga_brca_multi")
}

}
