% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.R
\name{tiny_perturb}
\alias{tiny_perturb}
\title{Load perturbation / CRISPR screen datasets}
\usage{
tiny_perturb(name = "k562_crispr", verbose = TRUE)
}
\arguments{
\item{name}{Dataset name}

\item{verbose}{Print loading info}
}
\value{
SingleCellExperiment or similar
}
\description{
Convenience function for perturbation data.
}
