% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.R
\name{tiny_scrna}
\alias{tiny_scrna}
\title{Load single-cell RNA-seq datasets}
\usage{
tiny_scrna(name = "pbmc", verbose = TRUE)
}
\arguments{
\item{name}{Dataset name: "zeisel", "darmanis", "pbmc", "baron", "muraro", "campbell"}

\item{verbose}{Print loading info}
}
\value{
SingleCellExperiment object
}
\description{
Convenience function for scRNA-seq data.
}
\examples{
\dontrun{
pbmc <- tiny_scrna("pbmc")
brain <- tiny_scrna("zeisel")
}

}
