% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.R
\name{tiny_spatial}
\alias{tiny_spatial}
\title{Load spatial transcriptomics datasets}
\usage{
tiny_spatial(name = "dlpfc", verbose = TRUE)
}
\arguments{
\item{name}{Dataset name: "dlpfc", "mouse_embryo", "brain_libd", "ovarian"}

\item{verbose}{Print loading info}
}
\value{
SpatialExperiment object
}
\description{
Convenience function for spatial data.
}
\examples{
\dontrun{
brain <- tiny_spatial("dlpfc")
embryo <- tiny_spatial("mouse_embryo")
}

}
