% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.R
\docType{package}
\name{tinyomics-package}
\alias{tinyomics}
\alias{tinyomics-package}
\title{tinyomics: Explore tiny omics datasets}
\description{
A lightweight interface to curated omics datasets from various Bioconductor packages. Provides easy access to single-cell RNA-seq, bulk RNA-seq, ATAC-seq, spatial transcriptomics, CyTOF, ChIP-seq, perturbation, genotype, and multi-omics datasets. All data returns as tidyomics-compatible container classes (SummarizedExperiment, SingleCellExperiment, SpatialExperiment, etc.). Dependencies are loaded dynamically to minimize package bloat.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tinyomics/tinyomics}
  \item Report bugs at \url{https://github.com/tinyomics/tinyomics/issues}
}

}
\author{
\strong{Maintainer}: Abdullah Al-Nahid \email{alnahid@usc.com}

}
\keyword{internal}
