% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.R
\name{tinyomics}
\alias{tinyomics}
\alias{scrna}
\alias{bulk}
\alias{spatial}
\alias{cytof}
\alias{chipseq}
\alias{atac}
\alias{perturb}
\alias{tiny_list}
\alias{tiny_search}
\title{tinyomics}
\description{
Unified access to omics datasets
}
